Create Checking Prime
to test whether the given number is prime number or not(Based on fact that one of the factors of the given no. is less than or equal to square root of the given no. and other is greater than or equal to square root of the given no. )
Time Complexity of this algorithm is O(sqrt(N))